requirements.txt
affine==2.4.0
aiobotocore==2.15.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.35.34
botocore==1.35.34
bounded-pool-executor==0.0.3
branca==0.8.0
Cartopy==0.23.0
certifi==2024.8.30
cffi==1.17.1
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dask==2023.7.0
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
earthaccess==0.5.2
earthpy==0.9.4
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
folium==0.17.0
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
geopandas==1.0.1
h11==0.14.0
h5netcdf==1.3.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
locket==1.0.0
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
multimethod==1.12
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.6.4
networkx==3.3
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.23.5
openai==0.2.0
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pqdm==0.2.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydap==3.4.0
Pygments==2.18.0
pyhdf==0.11.4
pyogrio==0.10.0
pyparsing==3.0.9
pyproj==3.7.0
pyshp==2.3.1
python-cmr==0.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.4.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rioxarray==0.17.0
rpds-py==0.20.0
s3fs==2023.12.2
s3transfer==0.10.2
scikit-image==0.24.0
scipy==1.14.1
Send2Trash==1.8.3
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
terminado==0.18.1
tifffile==2024.9.20
tinycss2==1.3.0
tinynetrc==1.3.1
tomli==2.0.2
toolz==1.0.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
WebOb==1.8.8
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.16.0
xarray==2023.7.0
xyzservices==2024.9.0
yarl==1.13.1
zipp==3.20.2
